Defect/json syntax exception

+++ blotoutSDK/src/main/java/com/analytics/blotout/data/database/EventDatabaseService.kt
@@ -1,15 +1,15 @@
 package com.analytics.blotout.data.database
 
 import android.util.Log
-import com.analytics.blotout.AnalyticsActivityLifecycleCallbacks
+import com.analytics.blotout.BlotoutAnalytics
 import com.analytics.blotout.DependencyInjectorImpl
 import com.analytics.blotout.data.database.entity.EventEntity
 import com.analytics.blotout.model.Events
 import com.analytics.blotout.network.ApiDataProvider
 import com.google.gson.Gson
+import com.google.gson.JsonSyntaxException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import retrofit2.Call
@@ -23,53 +23,65 @@ class EventDatabaseService {
     }
 
     fun insertEvent(event: EventEntity) {
-        try {
-            CoroutineScope(Dispatchers.Default).launch {
+
+        CoroutineScope(Dispatchers.Default).launch {
+            try {
                 DependencyInjectorImpl.getInstance().getEventDatabase().eventDao()
-                    .insertEvent(event)
+                        .insertEvent(event)
+            } catch (e: Exception) {
+                Log.e(TAG, e.toString())
             }
-        }catch (e:Exception){
-            Log.e(TAG,e.toString())
         }
     }
 
     fun getEvents() {
-        try {
-            CoroutineScope(Dispatchers.Default).launch {
-                val idTable = ArrayList<Int>()
-                evenDao.getEvents().collect { data ->
-                    data.forEach {
+
+        CoroutineScope(Dispatchers.Default).launch {
+
+            val idTable = ArrayList<Int>()
+            evenDao.getEvents().collect { data ->
+                data.forEach {
+                    try {
                         if (!idTable.contains(it.id)) {
                             idTable.add(it.id)
-                            var eventData = it.eventData
+                            val eventData = it.eventData
                             Log.d("###Pushing id ", "" + it.id)
                             Log.d("###Pushing events ", "" + eventData)
-                            var eventObject = Gson().fromJson(eventData, Events::class.java)
+
+                            val eventObject: Events = Gson().fromJson(eventData.trim(), Events::class.java)
+
                             DependencyInjectorImpl.getInstance().getConfigurationManager()
-                                .publishEvents(eventObject, object : ApiDataProvider<Any?>() {
-                                    override fun onFailed(
-                                        errorCode: Int,
-                                        message: String,
-                                        call: Call<Any?>
-                                    ) {
-                                    }
-
-                                    override fun onError(t: Throwable, call: Call<Any?>) {
-                                    }
-
-                                    override fun onSuccess(data: Any?) {
-                                        CoroutineScope(Dispatchers.Default).launch {
-                                            //Log.d("###deleting id ", "" + it.id)
-                                            evenDao.deleteEvent(it)
+                                    .publishEvents(eventObject, object : ApiDataProvider<Any?>() {
+                                        override fun onFailed(
+                                                errorCode: Int,
+                                                message: String,
+                                                call: Call<Any?>
+                                        ) {
                                         }
-                                    }
-                                })
+
+                                        override fun onError(t: Throwable, call: Call<Any?>) {
+                                        }
+
+                                        override fun onSuccess(data: Any?) {
+                                            CoroutineScope(Dispatchers.Default).launch {
+                                                //Log.d("###deleting id ", "" + it.id)
+                                                evenDao.deleteEvent(it)
+                                            }
+                                        }
+                                    })
+                        }
+                    } catch (e: Exception) {
+                        when (e) {
+                            is JsonSyntaxException, is IllegalStateException -> {
+                                val malformedLog = hashMapOf<String, Any>()
+                                malformedLog.put(it.id.toString(), it.eventData)
+                                BlotoutAnalytics.capture("Malformed Log", malformedLog)
+                                evenDao.deleteEvent(it)
+                            }
                         }
                     }
                 }
             }
-        }catch (e:Exception){
-            Log.e(TAG,e.toString())
         }
     }
 }
